add simple routing, so we can start connecting pages together
